move GB18030
